requirements.txt
absl-py==0.13.0
addict==2.4.0
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
charset-normalizer==2.0.4
chumpy==0.70
cloudpickle==2.0.0
cmake==3.21.4
cupy-cuda102==9.5.0
cycler==0.10.0
Cython==3.0.7
decorator==5.0.6
easydict==1.10
einops==0.3.2
fastrlock==0.8
filelock==3.12.2
fsspec==2023.1.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
h5py==3.4.0
huggingface-hub==0.16.4
idna==3.2
imageio==2.9.0
importlib-metadata==6.7.0
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.17.0
joblib==1.0.1
json-tricks==3.17.3
kiwisolver==1.3.1
lightrun==0.1.46
Markdown==3.3.4
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mmcv-full==1.3.8
mmpose==0.29.0
munkres==1.1.4
numexpr==2.7.3
numpy==1.21.6
oauthlib==3.1.1
olefile==0.46
opencv-python==4.9.0.80
packaging==23.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pip==21.0.1
platformdirs==4.0.0
prompt-toolkit==3.0.17
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.1
pykeops==1.5
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.2.1
setuptools==52.0.0.post20210125
six==1.15.0
table==0.0.4
tables==3.6.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
timm==0.6.13
tomli==2.0.1
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchpq==0.2.0.3
torchvision==0.8.2
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==4.7.1
urllib3==1.26.6
vispy==0.5.3
wcwidth==0.2.5
Werkzeug==2.0.1
wheel==0.36.2
xtcocotools==1.14.3
yapf==0.40.1
zipp==3.5.0
